requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
python-dotenv
psycopg2-binary
gunicorn
ALLOWED_ORIGINS=https://multywave.com,https://www.multywave.com
